% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedfileHG38.R
\name{bedfileHG38}
\alias{bedfileHG38}
\title{Load bed file}
\usage{
bedfileHG38()
}
\description{
Load a dataset with information about the 37 genes within the mitochondrial genome.
}
\examples{
 data("bedfileMito")
}
