% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateQCMetricsSeq.R
\name{calculateQCMetricsSeq}
\alias{calculateQCMetricsSeq}
\title{Calculate QC metrics from sequencing - FastQC}
\usage{
calculateQCMetricsSeq(patients_dir, trimmed = FALSE)
}
\arguments{
\item{patients_dir}{Path of the directory that contains R1 and R2 fastq files. It can be either the "trimmed" folder or the original folder.}

\item{trimmed}{kjvdfk}
}
\value{
scores_qc df with metrics from FastQC
}
\description{
Executes tkjvnfkd.
}
