% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateQCMetricsSTAR.R
\name{calculateQCMetricsSTAR}
\alias{calculateQCMetricsSTAR}
\title{Calculate QC metrics - STAR}
\usage{
calculateQCMetricsSTAR(patients_dir, trimmed = FALSE)
}
\arguments{
\item{patients_dir}{Path of the directory that contains R1 and R2 fastq files. It can be either the "trimmed" folder or the original folder.}

\item{trimmed}{kjvdfk}
}
\value{
metricasSTAR df with metrics from alignment
}
\description{
Executes tkjvnfkd.
}
