% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlSelection_cnv.R
\name{controlSelection_cnv}
\alias{controlSelection_cnv}
\title{Control Selection for CNVs}
\usage{
controlSelection_cnv(PatientsDB, bed)
}
\arguments{
\item{PatientsDB}{database in dataframe format that contains the counts of
the genes of any patient (normal or not) that has been analyzed.}

\item{bed}{is a dataframe that has start, end and name of exons within the panel selected.}

\item{id}{number or character that indicates which patient will be analyzed.
It has to be the same as its id in the data base.}
}
\value{
controles dataframe that indicates which samples has been selected as references and it frequency.
}
\description{
Using the ExomeDepth algorithm to select the patient that is more likely to be "normal" (no CNVs).
}
