% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSignatures.R
\name{Correlacion_Prolif}
\alias{Correlacion_Prolif}
\title{Run Gene Signatures to MX.}
\usage{
Correlacion_Prolif(cohorte, TPMt)
}
\arguments{
\item{patients_dir}{path to the folder that contains one folder for each sample.}

\item{Metadata}{dataframe that contains information about all the samples. One sample per row.}

\item{group}{must be a name of one of the columns of the metadata}
}
\value{
Todos_FusionReport and Stats_Fusions: 2 dataframes.
}
\description{
Evaluates specific genes for proliferation, ca20 and cytolicitc score.
}
\examples{
fusionStats(patients_dir, Metadata, group = "group")
}
