% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mycofree.R
\name{downloadBBmap}
\alias{downloadBBmap}
\title{downloadBBmap}
\usage{
downloadBBmap(path_for_BBMap)
}
\arguments{
\item{path_for_BBMap}{Path of the directory where the software will be stored.}
}
\value{
path_bbmap path of the software.
}
\description{
Installation of BBmap
}
