% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadMIXCR.R
\name{downloadMIXCR}
\alias{downloadMIXCR}
\title{downloadMIXCR}
\usage{
downloadMIXCR()
}
\value{
The path where the executable file is located
}
\description{
Downloads and decompresses the MIXCR software
}
