% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadMIXTURE.R
\name{downloadMIXTURE}
\alias{downloadMIXTURE}
\title{downloadHG38}
\usage{
downloadMIXTURE()
}
\value{
Paths of FASTA and GTF(annotation) from the genome reference.
}
\description{
Downloads the FASTA and the annotation of the genome reference version HG38.
}
