% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enriquecimiento_Vias.R
\name{Enriquecimiento_Vias}
\alias{Enriquecimiento_Vias}
\title{Enriquecimiento vias}
\usage{
Enriquecimiento_Vias(
  b_G,
  categoria = "MTT",
  Grupos,
  p = 0.05,
  lfc = 1,
  genes_seleccionados = NULL,
  listado = ""
)
}
\arguments{
\item{b_G}{Either the BAMfastq files}

\item{categoria}{Filters applied to the anded by GATK}

\item{Grupos}{visuf}

\item{p}{pvalue}

\item{lfc}{log fold change}

\item{genes_seleccionados}{list of kfernkdf name of the list}

\item{listado}{ofsijkkds}
}
\value{
vias egvfnjrnfk
}
\description{
It gein MitoR DataBase.
}
\examples{
genes_seleccionados <- Solicitud_2Br_DEG_en_G3vsG4$mergex3
vias <- Enriquecimiento_Vias(b_G = b_G, Grupos = "G18G19", genes_seleccionados = genes_seleccionados, listado = "mergex3")
}
