% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpresionDiferencial_Listados.R
\name{ExpresionDiferencial_Listados}
\alias{ExpresionDiferencial_Listados}
\title{Expresion Diferencial Listados}
\usage{
ExpresionDiferencial_Listados(
  b,
  df_listado,
  Grupos,
  tejido,
  lfc = 1,
  p = 0.05,
  categoria = "MTT",
  plotMDS = FALSE,
  plot_heatmaps = TRUE,
  porListadoGenes = TRUE
)
}
\arguments{
\item{b}{DGEList}
}
\value{
Control DB updated
}
\description{
Generates differential expression analysis.
Selects differential genes (depending on p value and lfc) and identify which of them are included in each column in df_listado.
Creates heatmaps for:
\itemize{
\item General DEG
\item Genes listed in df_listado
\item Genes DEG included in listes.
}
}
\examples{
ControlDB <- AddControl(path_dir ="/home/sam/Patients/123/123.fastq")
}
