% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCtoDGEList.R
\name{FCtoDGEList}
\alias{FCtoDGEList}
\title{Analyze copy number variations (CNVs) of a patient sample.}
\usage{
FCtoDGEList(path_fc, path_metadata)
}
\arguments{
\item{path_fc}{path to the feature counts rds file}

\item{path_metadata}{path to the metadata excel file}
}
\value{
a list:
}
\description{
Gjrdbhjdbj
}
