% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionStatsV.R
\name{fusionStatsV}
\alias{fusionStatsV}
\title{Obtain statistics from fusions.}
\usage{
fusionStatsV(
  patients_dir,
  Metadata = NA,
  group = NA,
  cohorte = "",
  sobrevida = TRUE
)
}
\arguments{
\item{patients_dir}{path to the folder that contains one folder for each sample.}

\item{Metadata}{dataframe that contains information about all the samples. One sample per row.}

\item{group}{must be a name of one of the columns of the metadata}
}
\value{
Todos_FusionReport and Stats_Fusions: 2 dataframes.
}
\description{
Generates 2 dataframes: 1. Rbind from every fusion report of the samples and 2. Stats from all samples indicating group.
}
\examples{
fusionStats(patients_dir, Metadata, group = "group")
}
