% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnv.R
\name{GetCNVsAnnotation}
\alias{GetCNVsAnnotation}
\title{Get CNVs annotation.}
\usage{
GetCNVsAnnotation(chr, x, countThreshold = 10)
}
\arguments{
\item{chr}{dataframe that contains the information about the CNVs detected.}

\item{countThreshold}{is a number}
}
\value{
lckerm wdkme
}
\description{
Generates a graph that represents the CNVs detected. The user will be able to
visualize the distribution, sizes, and positions of the variations within the whole mtDNA
and the genes that are affected by them.
}
