% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdownloadARRIBA.R
\name{IdownloadARRIBA}
\alias{IdownloadARRIBA}
\title{downloadArriba}
\usage{
IdownloadARRIBA()
}
\value{
The path where the .exe file is located
}
\description{
Downloads and decompresses the GATK software
}
