% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdownloadFastQC.R
\name{IdownloadFastQC}
\alias{IdownloadFastQC}
\title{downloadFastQC}
\usage{
IdownloadFastQC(
  soft_directory = sprintf("\%s/OMICsdoSof", Sys.getenv("R_LIBS_USER"))
)
}
\value{
The path where the .exe file is located
}
\description{
Downloads and decompresses the FastQC software
}
