% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdownloadSTAR.R
\name{IdownloadSTAR}
\alias{IdownloadSTAR}
\title{downloadSTAR}
\usage{
IdownloadSTAR(
  soft_directory = sprintf("\%s/OMICsdoSof", Sys.getenv("R_LIBS_USER"))
)
}
\value{
The path where the .exe file is located
}
\description{
Downloads and decompresses the GATK software
}
