% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdownloadTrimgalore.R
\name{IdownloadTrimGalore}
\alias{IdownloadTrimGalore}
\title{downloadTrimGalore}
\usage{
IdownloadTrimGalore()
}
\value{
The path where the .exe file is located
}
\description{
Downloads and decompresses the TrimGalore software
}
