% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertPathtoSoft.R
\name{insertPathtoSoft}
\alias{insertPathtoSoft}
\title{insert Path to Soft}
\usage{
insertPathtoSoft(software_name, path)
}
\arguments{
\item{software_name}{name of the software you want to save.}

\item{path}{path of the executable of the software already installed.}
}
\description{
It saves the path of a software in the path_to_soft.txt file in OMICs library.
}
