% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mycofree.R
\name{mapeoMycoplasma_Grupal}
\alias{mapeoMycoplasma_Grupal}
\title{mapeoMycoplasma_Grupal}
\usage{
mapeoMycoplasma_Grupal(
  dir_patients,
  index_rsubread,
  bbmap = FALSE,
  path_bbmap = NA,
  rate = "0.6"
)
}
\arguments{
\item{dir_patients}{Path where the patients are stored.}

\item{index_rsubread}{Path where the index of mycoplasma generated with Rsubread is stored.}

\item{bbmap}{TRUE/FALSE. Set to TRUE in order to subsample the fastq files.}

\item{path_bbmap}{Path where the software is stored. Set to NA if BBmap is not needed.}

\item{rate}{string from 0 to 1 indicating the percentage of the output fastq. For example: rate = "0.6" will generate fastq files with 60\% of the original size.}
}
\value{
mapeos_mycoplasma: df with percentage of mapped reads to mycoplasma per sample.
}
\description{
Execution of mapeoMycoplasma_Individual to every patient of the cohort.
}
