% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mycofree.R
\name{mapeoMycoplasma_Individual}
\alias{mapeoMycoplasma_Individual}
\title{mapeoMycoplasma_Individual}
\usage{
mapeoMycoplasma_Individual(
  patient_dir,
  index_rsubread,
  bbmap = FALSE,
  path_bbmap = NA,
  rate = "0.6"
)
}
\arguments{
\item{patient_dir}{Path where the patients are stored.}

\item{index_rsubread}{Path where the index of mycoplasma generated with Rsubread is stored.}

\item{bbmap}{TRUE/FALSE. Set to TRUE in order to subsample the fastq files.}

\item{path_bbmap}{Path where the software is stored. Set to NA if BBmap is not needed.}

\item{rate}{string from 0 to 1 indicating the percentage of the output fastq. For example: rate = "0.6" will generate fastq files with 60\% of the original size.}
}
\value{
mapeo_mycoplasma: df with ID and percentage of mapped reads to mycoplasma.
}
\description{
Execution of mapeoMycoplasma_Individual to every patient of the cohort.
}
