% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFastQC.R
\name{plotFastQC_PBSQ}
\alias{plotFastQC_PBSQ}
\title{plotFastQC_PBSQ}
\usage{
plotFastQC_PBSQ(patients_dir, trimmed = FALSE)
}
\arguments{
\item{patients_dir}{Path of the directory that contains one folder with fastq files (R1 and R2) of each patient.}

\item{trimmed}{it is set to TRUE when the fastQC files that you want to plot came from a trimmed file.}
}
\description{
generates the principal plot ("Per Base Sequence Quality") which compares the quality for all the samples.
}
