% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqP.R
\name{RNAseqP}
\alias{RNAseqP}
\title{RNAseqP}
\usage{
RNAseqP(
  patients_dir,
  genomeRef = "HG38",
  fastQC_after_trim = FALSE,
  plot_FastQC_trim = FALSE,
  trim_quality = 30,
  RunARRIBA = FALSE,
  RunFeatureCounts = TRUE,
  RunMIXCR = FALSE
)
}
\arguments{
\item{genomeRef}{indicated which reference will be used. Write "HG38" or ...}

\item{fastQC_after_trim}{lfnrks}

\item{RunARRIBA}{kvsfj}

\item{RunFeatureCounts}{vshbfj}

\item{RunMIXCR}{kvsjhbfj}

\item{patient_dir}{Path of the directory that contains R1 and R2 fastq files. It can be either the "trimmed" folder or the original folder.}
}
\value{
softwares that were runned and the time they took.
}
\description{
Executes the
}
