% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mycofree.R
\name{runBBmap}
\alias{runBBmap}
\title{runBBmap}
\usage{
runBBmap(path_bbmap, patient_dir, rate)
}
\arguments{
\item{path_bbmap}{Path where the software is stored.}

\item{patient_dir}{Path where the patient is stored.}

\item{rate}{string from 0 to 1 indicating the percentage of the output fastq. For example: rate = "0.6"}
}
\value{
list(R1_bbmap, R2_bbmap).
}
\description{
Execution of BBmap
}
