% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMIXCR.R
\name{runMIXCR}
\alias{runMIXCR}
\title{runMIXCR}
\usage{
runMIXCR(patient_dir)
}
\arguments{
\item{patient_dir}{the path of the folder´s patient yoy want to analyze. It can be the one that contains the R1 and R2 original or the trimmed files.}
}
\value{
The path where the executable file is located
}
\description{
Runs the MIXCR software
}
