% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMIXTURE.R
\name{runMIXTURE}
\alias{runMIXTURE}
\title{downloadHG38}
\usage{
runMIXTURE()
}
\value{
Paths of FASTA and GTF(annotation) from the genome reference.
}
\description{
Downloads the FASTA and the annotation of the genome reference version HG38.
}
