% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMitoRStats_HMTVAR.R
\name{update_freq_MitoR}
\alias{update_freq_MitoR}
\title{View Patient's Report}
\usage{
update_freq_MitoR(path_dir)
}
\value{
message that indicates that the report has been updated and downloaded.
}
\description{
Downloads the xlsx report of a specific patient with the data updated.
}
\examples{
view_patient()
}
